Replace fixed-width integer types with their fast variants when possible
On most systems, a type like uint8_t will optimize for memory, while a
type like uint_fast8_t will optimize for speed based on the compiler.
This seems to be the better choice considering how little memory is used
already.

+++ lz78-optimized/io.c
@@ -78,7 +78,7 @@ void buffer_pair(int fd, Code c, Symbol s, int width) {
     }
   }
 
-  for (int i = 0; i < BYTE; i += 1) {
+  for (uint_fast8_t i = 0; i < BYTE; i += 1) {
     if (s & (1 << i)) {
       bitbuf[bits / BYTE] |= (1 << (bits % BYTE));
     }
@@ -121,7 +121,7 @@ bool read_pair(int fd, Code *c, Symbol *s, int width) {
     bits = (bits + 1) % (BLOCK * BYTE);
   }
 
-  for (int i = 0; i < BYTE; i += 1) {
+  for (uint_fast8_t i = 0; i < BYTE; i += 1) {
     if (!bits) {
       read_bytes(fd, bitbuf, BLOCK);
     }
